requirements.txt
# Core dependencies
openai>=1.58.1
pydantic>=2.10.4
python-dotenv>=1.0.1
requests>=2.32.3

# Web search dependencies
google-api-python-client>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Streamlit and UI dependencies
streamlit>=1.41.1
streamlit-markmap>=1.0.1
streamlit-tags>=1.2.8

# Development dependencies
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# Optional: For better development experience
ipython>=8.0.0
jupyter>=1.0.0
